Refactor some of the code layout
